OPHSUPA-55: Refaktoroitu Opiskeluoikeudet erilliseen komponenttiin
