+++ suorituspalvelu-ui/src/components/HenkiloPage.tsx
@@ -1,121 +1,9 @@
 'use client';
-import { Opiskeluoikeus } from '@/api';
 import { FullSpinner } from '@/components/FullSpinner';
 import { useOppija } from '@/queries';
-import { Box, Stack } from '@mui/material';
-import {
-  ophColors,
-  OphLink,
-  OphTypography,
-} from '@opetushallitus/oph-design-system';
-import { Suspense, use } from 'react';
-import { PaperWithTopColor } from './PaperWithTopColor';
-import { useTranslate } from '@tolgee/react';
-import { LabeledInfoItem } from './LabeledInfoItem';
-import { Circle } from '@mui/icons-material';
-import { currentFinnishDate, isInRange } from '@/lib/time-utils';
-import { formatDate } from 'date-fns';
-import { NDASH } from '@/common';
-import { getOppilaitosLinkUrl } from '@/lib/getOppilaitosLink';
-import { configPromise } from '@/configuration';
-
-const VoimassaoloBadge = ({
-  voimassaolonAlku,
-  voimassaolonLoppu,
-}: {
-  voimassaolonAlku?: Date;
-  voimassaolonLoppu?: Date;
-}) => {
-  const { t } = useTranslate();
-
-  const isVoimassa = isInRange(
-    currentFinnishDate(),
-    voimassaolonAlku,
-    voimassaolonLoppu,
-  );
-
-  return (
-    <Stack sx={{ alignItems: 'center', flexDirection: 'row', gap: 2 }}>
-      <Box>
-        {voimassaolonAlku ? formatDate(voimassaolonAlku, 'd.M.y') : ''}
-        {` ${NDASH} `}
-        {voimassaolonLoppu ? formatDate(voimassaolonLoppu, 'd.M.y') : ''}
-      </Box>
-      <Stack sx={{ alignItems: 'center', flexDirection: 'row', gap: 0.5 }}>
-        <Circle
-          sx={{
-            fontSize: '18px',
-            color: isVoimassa ? ophColors.green3 : ophColors.orange3,
-          }}
-        />
-        <OphTypography>
-          {isVoimassa ? t('oppija.voimassa') : t('oppija.ei-voimassa')}
-        </OphTypography>
-      </Stack>
-    </Stack>
-  );
-};
-
-const Opiskeluoikeudet = ({
-  opiskeluoikeudet,
-}: {
-  opiskeluoikeudet: Array<Opiskeluoikeus>;
-}) => {
-  const { t } = useTranslate();
-  const config = use(configPromise);
-  return (
-    <Box data-test-id="opiskeluoikeudet">
-      <OphTypography variant="h2" component="h5" sx={{ marginBottom: 2 }}>
-        {t('oppija.opiskeluoikeudet')}
-      </OphTypography>
-      {opiskeluoikeudet.map((oo) => {
-        return (
-          <PaperWithTopColor
-            key={`${oo.tutkinto}-${oo.oppilaitos.oid}`}
-            topColor={ophColors.red1}
-          >
-            <Stack
-              direction="column"
-              gap={1}
-              data-test-id="opiskeluoikeus-paper"
-            >
-              <OphTypography variant="label" sx={{ marginBottom: 1 }}>
-                {oo.tutkinto}
-              </OphTypography>
-              <Stack
-                direction="row"
-                gap={1}
-                sx={{
-                  '& > *': { flex: '1 1 auto' },
-                }}
-              >
-                <LabeledInfoItem
-                  label={t('oppija.oppilaitos')}
-                  value={
-                    <OphLink
-                      href={getOppilaitosLinkUrl(config, oo.oppilaitos.oid)}
-                    >
-                      {oo.oppilaitos.nimi}
-                    </OphLink>
-                  }
-                />
-                <LabeledInfoItem
-                  label={t('oppija.voimassaolo')}
-                  value={
-                    <VoimassaoloBadge
-                      voimassaolonAlku={oo.voimassaolonAlku}
-                      voimassaolonLoppu={oo.voimassaolonLoppu}
-                    />
-                  }
-                />
-              </Stack>
-            </Stack>
-          </PaperWithTopColor>
-        );
-      })}
-    </Box>
-  );
-};
+import { Stack } from '@mui/material';
+import { Suspense } from 'react';
+import { Opiskeluoikeudet } from './Opiskeluoikeudet';
 
 const OppijaContent = ({ oppijaNumero }: { oppijaNumero: string }) => {
   const { data: tiedot } = useOppija(oppijaNumero);
